requirements.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
chardet==5.2.0 ; python_version >= "3.12" and python_version < "4"
charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
config==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
crispy-bootstrap4==2024.1 ; python_version >= "3.12" and python_version < "4.0"
crispy-tailwind==1.0.3 ; python_version >= "3.12" and python_version < "4.0"
django-allauth==0.63.6 ; python_version >= "3.12" and python_version < "4.0"
django-appconf==1.0.6 ; python_version >= "3.12" and python_version < "4.0"
django-browser-reload==1.13.0 ; python_version >= "3.12" and python_version < "4.0"
django-ckeditor==6.7.1 ; python_version >= "3.12" and python_version < "4.0"
django-colorfield==0.11.0 ; python_version >= "3.12" and python_version < "4.0"
django-compressor==4.5 ; python_version >= "3.12" and python_version < "4.0"
django-crispy-forms==2.3 ; python_version >= "3.12" and python_version < "4.0"
django-environ==0.11.2 ; python_version >= "3.12" and python_version < "4"
django-filter==24.2 ; python_version >= "3.12" and python_version < "4.0"
django-formtools==2.5.1 ; python_version >= "3.12" and python_version < "4.0"
django-image-cropping==1.7 ; python_version >= "3.12" and python_version < "4.0"
django-imagekit==5.0.0 ; python_version >= "3.12" and python_version < "4.0"
django-js-asset==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
django-mathfilters==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
django-pictures==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
django-stdimage==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
django-tailwind==3.8.0 ; python_version >= "3.12" and python_version < "4.0"
django-widget-tweaks==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
django==5.0.6 ; python_version >= "3.12" and python_version < "4.0"
easy-thumbnails==2.10 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.23.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.36.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-secret-manager==2.21.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.13.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.67.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.67.1 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
pilkit==3.0 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.3.0 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.25.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.3 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
rcssmin==1.1.2 ; python_version >= "3.12" and python_version < "4.0"
reportlab==4.2.0 ; python_version >= "3.12" and python_version < "4"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rjsmin==1.2.2 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
sqlparse==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
